refactor: improve root command name
