+++ src/main/kotlin/Main.kt
@@ -1,5 +1,3 @@
 package com.neo.envmanager
 
-import com.neo.envmanager.core.Start
-
-fun main(args: Array<String>) = Start().main(args)
+fun main(args: Array<String>) = Envm().main(args)
